osa4

+++ osa4/blogilista/controllers/blogs.js
@@ -1,13 +1,31 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const jwt = require('jsonwebtoken')
+const { tokenExtractor, userExtractor } = require('../utils/middleware')
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog.
+    find({}).populate('user')
   response.json(blogs)
 })
 
-blogsRouter.post('/', async (request, response) => {
-  const blog = new Blog(request.body)
+blogsRouter.post('/', tokenExtractor, userExtractor, async (request, response) => {
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+
+  const user = request.user
+
+  if (!user) {
+    return response.status(400).json({ error: 'userId missing or not valid' })
+  }
+
+  const blog = new Blog({
+    ...request.body,
+    user: user._id
+  })
 
   if (!blog.likes) {
     blog.likes = 0
@@ -17,11 +35,46 @@ blogsRouter.post('/', async (request, response) => {
   }
 
   const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+
   response.status(201).json(savedBlog)
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
+blogsRouter.delete('/:id', tokenExtractor, async (request, response) => {
+  const blog = await Blog.findById(request.params.id)
+  if (!blog) {
+    return response.status(404).json({ error: 'blog not found' })
+  }
+
+  if (!request.token) {
+    return response.status(401).json({ error: 'token missing' })
+  }
+
+  let decodedToken
+  try {
+    decodedToken = jwt.verify(request.token, process.env.SECRET)
+  } catch {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+
+  const user = request.user
+  if (!user) {
+    return response.status(401).json({ error: 'user not found' })
+  }
+
+  if (blog.user && blog.user.toString() !== user.id.toString()) {
+    return response.status(403).json({ error: 'user must be the one who added the blog' })
+  }
+
   await Blog.findByIdAndDelete(request.params.id)
+
+  user.blogs = user.blogs.filter(b => b.toString() !== blog._id.toString())
+  await user.save()
+
   response.status(204).end()
 })
